+++ server/app/enums/__init__.py
@@ -1,0 +1,10 @@
+from app.enums.clothes import (
+    CategoryEnum,
+    SizeEnum,
+    StatusEnum,
+    StyleEnum,
+    SeasonEnum,
+    NoteEnum,
+    ColorEnum,
+    MaterialsEnum
+)
